% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateList.R
\name{annotateList}
\alias{annotateList}
\title{Annotate a List of Gene Markers with Multiple Databases}
\usage{
annotateList(input_list)
}
\arguments{
\item{input_list}{A list of gene markers.}
}
\value{
A data frame containing the annotated gene markers with consensus cell types.
}
\description{
This function takes a list of gene markers and annotates them using BRETIGEA,
CellMarker 2.0, PanglaoDB, and Human Protein Atlas databases.
}
