% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/braincellann-package.R
\docType{package}
\name{braincellann-package}
\alias{braincellann}
\alias{braincellann-package}
\title{braincellann Package}
\description{
This package provides tools for annotating brain cell types using various biological databases.
}
\details{
#' @name braincellann
}
\keyword{internal}
